requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.24.0
motor==3.3.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
